Separate cloud function for account deletion into (1) auth user trigger + (2) user doc trigger, and refactor admin account deletion.

+++ functions/deletes.js
@@ -4,40 +4,45 @@ const admin = require('firebase-admin');
 const db = admin.firestore();
 const bucket = admin.storage().bucket();
 
-// remove all application docs and storage files for deleted user;
 exports.cleanupUserDelete = functions.auth.user()
-  .onDelete(async user => {
-    const uid = user.uid;
+  .onDelete(async user => await db.collection('users').doc(user.uid).delete());
 
-    // user document
-    const deleteUserDoc = await db.collection('users').doc(uid).delete();
+exports.cleanupUserContent = functions.firestore
+  .document('users/{userID}')
+  .onDelete(async (snap, context) => {
+    const uid = context.params.userID;
+    const listOfDeletions = [];
+    
+    // authentication account
+    const authUserExists = await admin.auth().getUser(uid);
+    if (authUserExists) listOfDeletions.push(await admin.auth().deleteUser(uid));
 
     // school application documents
     const querySchoolAppls = await db.collection('schoolApplications').where('authorID', '==', uid);
-    const [schoolFiles] = await bucket.getFiles({prefix: '/schools', delimiter: '/'}).catch(console.error);
-    const deleteSchoolDocs = deleteFirestoreDocs(querySchoolAppls);
-    const deleteSchoolFiles = deleteStorageFiles(schoolFiles, uid);
+    const [schoolFiles] = await bucket.getFiles({prefix: '/schools', delimiter: '/'});
+    listOfDeletions.push(deleteFirestoreDocs(querySchoolAppls));
+    listOfDeletions.push(deleteStorageFiles(schoolFiles, uid));
 
     // homestay application documents
     const queryHomestayAppls = await db.collection('homestayApplications').where('authorID', '==', uid);
-    const [homestayFiles] = await bucket.getFiles({prefix: '/homestays', delimiter: '/'}).catch(console.error);
-    const deleteHomestayDocs = deleteFirestoreDocs(queryHomestayAppls);
-    const deleteHomestayFiles = deleteStorageFiles(homestayFiles, uid);
+    const [homestayFiles] = await bucket.getFiles({prefix: '/homestays', delimiter: '/'});
+    listOfDeletions.push(deleteFirestoreDocs(queryHomestayAppls));
+    listOfDeletions.push(deleteStorageFiles(homestayFiles, uid));
 
     // airport application documents
     const queryAirportAppls = await db.collection('airportRideApplications').where('authorID', '==', uid);
-    const [airportFiles] = await bucket.getFiles({prefix: '/airportRides', delimiter: '/'}).catch(console.error);
-    const deleteAirportDocs = deleteFirestoreDocs(queryAirportAppls);
-    const deleteAirportFiles = deleteStorageFiles(airportFiles, uid);
+    const [airportFiles] = await bucket.getFiles({prefix: '/airportRides', delimiter: '/'});
+    listOfDeletions.push(deleteFirestoreDocs(queryAirportAppls));
+    listOfDeletions.push(deleteStorageFiles(airportFiles, uid));
     
-    return await Promise.all([deleteUserDoc, deleteSchoolDocs, deleteHomestayDocs, deleteAirportDocs, deleteSchoolFiles, deleteHomestayFiles, deleteAirportFiles]);
+    return await Promise.all(listOfDeletions);
 });
 
 function deleteFirestoreDocs(query) {
   const batch = db.batch();
   query.get().then(async snapshot => {
     if (snapshot.empty) {
-      console.log('No matching documents in Firestore DB found for this user id.');
+      console.log('No matching document(s) found from Firestore DB.');
       return;
     }
     snapshot.forEach(doc => batch.delete(doc.ref));
@@ -46,15 +51,15 @@ function deleteFirestoreDocs(query) {
 }
 
 function deleteStorageFiles(files, uid) {
+  if (!files.length) {
+    console.log("No matching storage file(s) found in Firebase Storage");
+    return;
+  }
   files.forEach(async file => {
     const [metadata] = await bucket.file(file.name).getMetadata();
-    
     if (metadata.metadata.owner === uid) {
       await bucket.file(file.name).delete();
       console.log(`Deleted ${file.name} from Firebase Storage.`);
     }
-
-    console.log('No matching storage file found for this uid.')
-    return;
   });
 }
